requirements.txt
django==3.2.3
djangorestframework==3.12.4